move otaclient._utils.typing into otaclient_common package
