+++ src/otaclient/configs/ecu_info.py
@@ -26,7 +26,7 @@ import yaml
 from pydantic import AfterValidator, BeforeValidator, Field, IPvAnyAddress
 from typing_extensions import Annotated
 
-from otaclient._utils.typing import NetworkPort, StrOrPath, gen_strenum_validator
+from otaclient_common.typing import NetworkPort, StrOrPath, gen_strenum_validator
 from otaclient.configs._common import BaseFixedConfig
 
 logger = logging.getLogger(__name__)
