+++ src/otaclient/configs/proxy_info.py
@@ -26,7 +26,7 @@ import yaml
 from pydantic import AliasChoices, AnyHttpUrl, Field, IPvAnyAddress
 from pydantic_core import Url
 
-from otaclient._utils.typing import NetworkPort, StrOrPath
+from otaclient_common.typing import NetworkPort, StrOrPath
 from otaclient.configs._common import BaseFixedConfig
 
 logger = logging.getLogger(__name__)
